Move FakeOutput resolving to back phase
